requirements.txt
# Python 3.12 / 3.14
# Core
langgraph>=0.2.55
langchain>=0.3.7
langchain-openai>=0.2.8
langchain-core>=0.3.15
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
PyYAML>=6.0.2

# Data / ML (NumPy-only model — no sklearn/scipy to avoid C build issues)
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.4.2

# HTTP / retry
httpx>=0.27.0
tenacity>=8.5.0

# UI
streamlit>=1.39.0

# Policy PDF generation
reportlab>=4.2.0

# Testing
pytest>=8.3.0

# PDF text extraction (pure-Python, no native deps)
pypdf>=4.0.0

# For full OCR: pip install -r requirements-ocr.txt (unstructured[pdf])
